refactor expected results utils into test_utils
functions are needed in all tests
